Add cross-platform JavaFX dependencies
Updated the JavaFX version to the latest version of JavaFX 17, which is
JavaFX 17.0.11. This version natively supports Apple Silicon.

The plugin used to download the JavaFX added in a previous commit seems
to only add the JavaFX version for the current operating system. Testing
the JAR created across multiple operating systems does not work since
the JavaFX runtime for the operating system required may not be present
in the fat JAR file generated.

This commit removes the plugin and replaces it by adding all versions of
JavaFX 17.0.11 for all operating systems and architectures offered
(windows, mac, mac-aarch64, linux) as a dependency, ensuring JAR files
work for all operating systems.