Swap Apple Silicon and x64 JavaFX imports
Putting mac-aarch64 after mac allows a JAR to be created on Macs with
Apple Silicon, but does not allow JARs created with other operating
systems to run on Apple Silicon.

However, due to this ordering change, there is a need to check whether
JARs created on other operating systems work on Intel Macs (since the
same issue might occur for Intel Macs).